+++ changelog.md
@@ -18,6 +18,8 @@ Codion Change Log
 - LocalEntityConnection.iterator() removed.
 ### is.codion.framework.db.rmi
 - DefaultRemoteEntityConnectionProvider now handles iterator() with RemoteEntityResultIteratorWrapper.
+### is.codion.framework.model
+- EntityEditModel.EditorValue.validate() added.
 ### is.codion.framework.server
 - DefaultRemoteEntityConnection, EntityConnection.iterator() implemented with timeouts.
 ### is.codion.swing.framework.ui